refactoring

+++ domain/unit_progress.go
@@ -3,11 +3,11 @@ package domain
 import "fmt"
 
 type UnitProgress struct {
-	Level      float32 `json:"level"`
-	Experience float32 `json:"experience"`
+	Level      uint `json:"level"`
+	Experience uint `json:"experience"`
 }
 
 func (p UnitProgress) String() string {
-	return fmt.Sprintf("level: %g, experience: %g",
+	return fmt.Sprintf("level: %d, experience: %d",
 		p.Level, p.Experience)
 }
